+++ src/main/java/info/coremodding/craftenchants/enchants/item/recipes/EnchantSword.java
@@ -3,7 +3,7 @@
   * http://www.gnu.org/licenses/gpl.html
   * http://www.gnu.org/licenses/lgpl.html
   */
-package info.coremodding.craftenchants.crafting;
+package info.coremodding.craftenchants.enchants.item.recipes;
 
 import info.coremodding.craftenchants.enchants.ItemCE;
 import info.coremodding.craftenchants.enchants.item.enchants.SharpeningStone;
