+++ src/main/java/info/coremodding/craftenchants/enchants/item/recipes/VanillaRecipes.java
@@ -2,7 +2,7 @@
  * Copyright 2014 by CoreModding under GNU Lesser General Public License (LGPLv3)
  * http://www.gnu.org/licenses/gpl.html http://www.gnu.org/licenses/lgpl.html
  */
-package info.coremodding.craftenchants.crafting;
+package info.coremodding.craftenchants.enchants.item.recipes;
 
 import net.minecraft.item.ItemStack;
 import cpw.mods.fml.common.registry.GameRegistry;
