+++ src/main/java/info/coremodding/craftenchants/handler/Crafting.java
@@ -5,10 +5,11 @@
   */
 package info.coremodding.craftenchants.handler;
 
-import info.coremodding.craftenchants.crafting.EnchantArmor;
-import info.coremodding.craftenchants.crafting.EnchantBow;
-import info.coremodding.craftenchants.crafting.EnchantSword;
-import info.coremodding.craftenchants.crafting.VanillaRecipes;
+import info.coremodding.craftenchants.enchants.item.recipes.EnchantArmor;
+import info.coremodding.craftenchants.enchants.item.recipes.EnchantBow;
+import info.coremodding.craftenchants.enchants.item.recipes.EnchantSword;
+import info.coremodding.craftenchants.enchants.item.recipes.VanillaRecipes;
+
 import cpw.mods.fml.common.registry.GameRegistry;
 
 public class Crafting {
